refactor(gateway): split GatewayInboundMixin into a facade + three topic siblings
gateway/run_inbound.py held one 2,023-line class (80 methods), past the ~2,000-line
split threshold in AGENTS.md. Mechanical extraction, no rewrites:

  gateway/run_inbound.py          2061 -> 698 lines  (facade; 26 methods + class-level state)
  gateway/run_inbound_hm.py      new, 1140 lines     (46 _hm* methods)
  gateway/run_inbound_enrich.py  new,  154 lines     (4 _enrich* methods)
  gateway/run_inbound_prepend.py new,   97 lines     (4 _prepend* methods)

Evidence: 80/80 methods byte-identical (AST source-segment equality vs pre-split),
py_compile + import smoke green, MRO verified, 35-file scoped suite at the same
423/4 as the untouched baseline (4 pre-existing failures, unrelated).
